Migrate to java.io.Serializable - manually implemented serializer to workaround BlockPos not implementing Serialzable
